refactor(ui): 重构批改界面并移除调试日志
- 将 GradeActivity.vue 重构为简单页面，提供返回旧版批改功能的链接
- 移除所有组件中的 console.error 和 console.log 调试输出
- 统一错误处理，将错误信息直接显示给用户
- 修复数据导入时的验证逻辑，允许缺少 prompts 字段
- 在 ProvidersSettings 中使用统一的 ElMessage 组件
- 添加 GradeLegacy 路由用于访问旧版批改功能
- 优化错误提示信息，包含具体错误详情
